Refactoring BasicOrchardFeaturesTesting
