--- Lombiq.Tests.UI.BasicOrchardFeaturesTesting/Lombiq.Tests.UI.BasicOrchardFeaturesTesting.csproj
@@ -1,39 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <IsPublishable>false</IsPublishable>
-    <DefaultItemExcludes>$(DefaultItemExcludes);.git*;node_modules\**</DefaultItemExcludes>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Title>Lombiq UI Testing Toolbox - Basic Orchard Features Testing</Title>
-    <Authors>Lombiq Technologies</Authors>
-    <Copyright>Copyright © 2020, Lombiq Technologies Ltd.</Copyright>
-    <Description>Lombiq UI Testing Toolbox - Basic Orchard Features Testing: Orchard Core feature UI testing-related extensions for the web app under test. See the project website for detailed documentation.</Description>
-    <PackageIcon>NuGetIcon.png</PackageIcon>
-    <PackageTags>OrchardCore;Lombiq;AspNetCore;Selenium;Atata;Shouldly;xUnit;Axe;AccessibilityTesting;UITesting;Testing;Automation</PackageTags>
-    <RepositoryUrl>https://github.com/Lombiq/UI-Testing-Toolbox</RepositoryUrl>
-    <PackageProjectUrl></PackageProjectUrl>
-    <PackageLicenseFile>License.md</PackageLicenseFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="License.md" Pack="true" PackagePath="" />
-    <None Include="Readme.md" />
-    <None Include="NuGetIcon.png" Pack="true" PackagePath="" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="node_modules\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Lombiq.Tests.UI\Lombiq.Tests.UI.csproj" />
-  </ItemGroup>
-
-</Project>
