--- Lombiq.Tests.UI.BasicOrchardFeaturesTesting/Readme.md
@@ -1,7 +1,0 @@
-# Lombiq UI Testing Toolbox - Basic Orchard Features Testing
-
-Orchard Core feature UI testing-related extensions for the web app under test.
-
-For general details about and on using the Toolbox see the [root Readme](../Readme.md).
-
-Note that the module depends on [Helpful Libraries](https://github.com/Lombiq/Helpful-Libraries).
