+++ Lombiq.Tests.UI/BasicOrchardFeaturesTesting/OrchardMediaOperationsTestingUITestContextExtensions.cs
@@ -5,7 +5,7 @@ using Lombiq.Tests.UI.Services;
 using OpenQA.Selenium;
 using System.Threading.Tasks;
 
-namespace Lombiq.Tests.UI.BasicOrchardFeaturesTesting.Extensions;
+namespace Lombiq.Tests.UI.BasicOrchardFeaturesTesting;
 
 /// <summary>
 /// Provides a set of extension methods for testing Orchard Core media operations.
